#46 fix usecase
